Implement stock stuff. Refactored product service

+++ src/main/java/com/ivomtdias/springshopapi/service/impl/ProductServiceImpl.java
@@ -2,12 +2,15 @@ package com.ivomtdias.springshopapi.service.impl;
 
 import com.ivomtdias.springshopapi.exception.ProductAlreadyExistsException;
 import com.ivomtdias.springshopapi.exception.ProductNotFoundException;
+import com.ivomtdias.springshopapi.exception.StockNotFoundException;
 import com.ivomtdias.springshopapi.model.Product;
+import com.ivomtdias.springshopapi.model.Stock;
 import com.ivomtdias.springshopapi.model.dto.ProductDTO;
 import com.ivomtdias.springshopapi.model.mapper.ProductDTOMapper;
 import com.ivomtdias.springshopapi.model.request.CreateProductRequest;
 import com.ivomtdias.springshopapi.model.request.UpdateProductRequest;
 import com.ivomtdias.springshopapi.repository.ProductRepository;
+import com.ivomtdias.springshopapi.repository.StockRepository;
 import com.ivomtdias.springshopapi.service.ProductService;
 import org.springframework.stereotype.Service;
 
@@ -22,10 +25,12 @@ public class ProductServiceImpl implements ProductService {
 
     private final ProductRepository productRepository;
     private final ProductDTOMapper productDTOMapper;
+    private final StockRepository stockRepository;
 
-    public ProductServiceImpl(ProductRepository productRepository, ProductDTOMapper productDTOMapper) {
+    public ProductServiceImpl(ProductRepository productRepository, ProductDTOMapper productDTOMapper, StockRepository stockRepository) {
         this.productRepository = productRepository;
         this.productDTOMapper = productDTOMapper;
+        this.stockRepository = stockRepository;
     }
 
     @Override
@@ -36,12 +41,16 @@ public class ProductServiceImpl implements ProductService {
                             throw new ProductAlreadyExistsException(createProductRequest.getName());
                         });
 
+        Product product = Product.builder()
+                .name(createProductRequest.getName())
+                .price(createProductRequest.getPrice())
+                .build();
+
+        stockRepository.save(new Stock(product, 0));
+
         return productDTOMapper.apply(
                 productRepository.save(
-                        Product.builder()
-                                .name(createProductRequest.getName())
-                                .price(createProductRequest.getPrice())
-                                .build()
+                        product
                 )
         );
     }
@@ -83,6 +92,12 @@ public class ProductServiceImpl implements ProductService {
     public void deleteProductById(final UUID productId) {
         if(productRepository.findById(productId).isEmpty())
             throw new ProductNotFoundException(productId);
+
+        Optional<Stock> stock = stockRepository.findStockById(productId);
+        if(stock.isEmpty())
+            throw new StockNotFoundException(productId);
+
+        stockRepository.deleteById(stock.get().getId());
 
         productRepository.deleteById(productId);
     }
